Add intro section to webpage
The webpage had a header, but no intro section to describe
the website. The intro section has been added and includes a
brief description of the website and a sign up button. There
is also a placeholder for a picture. The structure is created
using flex and the section is styled similar to the header.