Format and style call to action section
The call to action section gives the user one more way to
access the sign up button and encourage them to use the site.
This section has been added, formatted with flexbox, and
styled similarly to the rest of the webpage. It also includes
large margins and padding to ensure the webpage is not too
crowded.