Improve flow of website intro section
Before, when the webapge was downsized, the image in the intro
went off the side of the page if the display got too small.
Now, the flow has been improved so that the image wraps onto a
new row if the webapge gets too small so that the image is still
visible at smaller screen sizes. The text in the intro still
grows and shrinks dynamically while the image does not grow or
shrink.